requirements.txt
pandas>=2.0
numpy>=1.24
pandas-ta>=0.3.14b
yfinance>=0.2.31
pyyaml>=6.0
matplotlib>=3.7
tabulate>=0.9